Rename package gh_control -> ghcontrol
