+++ sourcetool/pkg/ghcontrol/connection_test.go
@@ -1,4 +1,4 @@
-package gh_control
+package ghcontrol
 
 import (
 	"testing"
