+++ sourcetool/pkg/ghcontrol/git_types.go
@@ -1,4 +1,4 @@
-package gh_control
+package ghcontrol
 
 import (
 	"fmt"
