+++ sourcetool/pkg/ghcontrol/notes_test.go
@@ -1,4 +1,4 @@
-package gh_control
+package ghcontrol
 
 import (
 	"context"
