+++ sourcetool/pkg/ghcontrol/options.go
@@ -1,4 +1,4 @@
-package gh_control
+package ghcontrol
 
 var defaultOptions = Options{
 	AllowMergeCommits: false,
